Implement synchronous AMQP channel.
A new boolean field 'is_sync" is added to the AMQP channel structure.
When active the channel will invoke the on-consume service before
sending an ACK to the queue. If the invoked service does not reply
with a 200 OK, the message will be rejected via a NACK with the
requeue flag set.

After the NACK the thread is put to sleep for a specific amount of
seconds to avoid the same message to be processed instantly (in case
of empty queue). In an upcoming commit the amount of seconds to
wait after a NACK will also become a parameter of the AMQP channel
entity.