Re-establish connection to AMQP when lost or simply shutdown from server.
Dropped connection are detected by pika both on a connection and on a
channel level. Proper callbacks are implemented to handle both scenarios.

BaseConnection reconnect timer is used during the reconnect procedure while
a (fixed) timeout of 30 seconds must expired before initiating the reconnection
procedure when the connection is lost with the server.